Added position to model and refactored methods

+++ src/modules/boards/boards.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { Injectable, NotFoundException } from '@nestjs/common';
 import { BoardsRepository } from './boards.repository';
 
 @Injectable()
@@ -9,19 +9,67 @@ export class BoardsService {
     return this.boardsRepository.findAll(userId);
   }
 
-  async findBoardById(id: string) {
-    return this.boardsRepository.findOne(id);
+  async findBoardById(id: string, userId: string) {
+    const board = await this.boardsRepository.findOneByUserAndId(id, userId);
+
+    if (!board) {
+      throw new NotFoundException('Board not found');
+    }
+
+    return board;
+  }
+
+  async getAllBoardsWithColumns(
+    userId: string,
+    position: 'asc' | 'desc' = 'asc',
+  ) {
+    return this.boardsRepository.findAllWithColumns(userId, position);
+  }
+
+  async findBoardByIdWithColumns(
+    id: string,
+    userId: string,
+    position: 'asc' | 'desc' = 'asc',
+  ) {
+    const board = await this.boardsRepository.findOneByUserAndIdWithColumns(
+      id,
+      userId,
+      position,
+    );
+
+    if (!board) {
+      throw new NotFoundException('Board not found');
+    }
+
+    return board;
   }
 
   async createBoard(title: string, description: string, userId: string) {
     return this.boardsRepository.create({ title, description, userId });
   }
 
-  async updateBoard(id: string, title?: string, description?: string) {
+  async updateBoard(
+    id: string,
+    userId: string,
+    title?: string,
+    description?: string,
+  ) {
+    const board = await this.boardsRepository.findOneByUserAndId(id, userId);
+
+    if (!board) {
+      throw new NotFoundException('Board not found');
+    }
+
     return this.boardsRepository.update(id, { title, description });
   }
 
-  async deleteBoard(id: string) {
+  async deleteBoard(id: string, userId: string) {
+    const board = await this.boardsRepository.findOneByUserAndId(id, userId);
+
+    if (!board) {
+      throw new NotFoundException('Board not found');
+    }
+
     return this.boardsRepository.delete(id);
   }
 }
